ONEM-28398 fix issues after uninstalling app with uninstallType=upgrade

+++ LISA/Executor.cpp
@@ -216,7 +216,22 @@ uint32_t Executor::Uninstall(const std::string& type,
         return ERROR_WRONG_PARAMS;
     }
 
-    if (! isAppInstalled(type, id, version)) {
+    // If an app was uninstalled earlier with uninstallType=upgrade, then the
+    // app record will still be inside the database. Also the data storage dir will
+    // still exist. Allow the uninstallation of these artifacts with "if test" below.
+    // Second "if test" is the uninstallation of the usual case: uninstalling an app
+    // of specific version.
+    if (version.empty() && !type.empty() && !id.empty() && uninstallType == "full") {
+        // verify that such an app record exists
+        if (dataBase->GetDataPaths(type, id).size() == 0) {
+            return ERROR_WRONG_PARAMS;
+        }
+        // only allowed when no specific version of app installed anymore
+        // if there are: the usual uninstall with a specific version should be called
+        if (dataBase->GetAppsPaths(type, id, "").size() > 0) {
+            return ERROR_WRONG_PARAMS;
+        }
+    } else if (!isAppInstalled(type, id, version)) {
         handle = "WrongParams";
         return ERROR_WRONG_PARAMS;
     }
@@ -343,39 +358,36 @@ uint32_t Executor::GetProgress(const std::string& handle, std::uint32_t& progres
     }
 }
 
-bool Executor::getStorageParamsValid(const std::string& type,
-        const std::string& id,
-        const std::string& version) const
-{
-    // In Stage 1 we support only none parameters or all of them
-    return ((!type.empty() && !id.empty() && !version.empty()) || (type.empty() && id.empty() && version.empty()));
-}
-
 uint32_t Executor::GetStorageDetails(const std::string& type,
                            const std::string& id,
                            const std::string& version,
                            Filesystem::StorageDetails& details)
 {
     namespace fs = Filesystem;
-    if(!getStorageParamsValid(type, id, version)) {
-        return ERROR_WRONG_PARAMS;
-    }
+
     try {
-        if(type.empty()) {
+        // if all params are empty then the overall disk usage is calculated
+        if(type.empty() && id.empty() && version.empty()) {
             INFO("Calculating overall usage");
             details.appPath = config.getAppsPath();
             details.appUsedKB = std::to_string((fs::getDirectorySpace(config.getAppsPath()) + fs::getDirectorySpace(config.getAppsTmpPath())) / 1024);
             details.persistentPath = config.getAppsStoragePath();
             details.persistentUsedKB = std::to_string(fs::getDirectorySpace(config.getAppsStoragePath()) / 1024);
-        } else {
+        } else if (!id.empty()) {
+            // When specific id is passed, calculate disk usage for this app.
+            // Type is optional since id is unique and sufficient. But if passed, it must match.
+            // If version is also passed: calculate the app size for this version, else not reported.
+            // Data storage size is always reported because this is version independent
             INFO("Calculating usage for: type = ", type, " id = ", id, " version = ", version);
-            std::vector<std::string> appsPaths = dataBase->GetAppsPaths(type, id, version);
-            unsigned long long appUsedKB{};
-            // In Stage 1 there will be only one entry here
-            for(const auto& i: appsPaths)
-            {
-                details.appPath = config.getAppsPath() + i;
-                appUsedKB += fs::getDirectorySpace(details.appPath);
+            if (!version.empty()) {
+                std::vector<std::string> appsPaths = dataBase->GetAppsPaths(type, id, version);
+                unsigned long long appUsedKB{};
+                // In Stage 1 there will be only one entry here
+                for (const auto &i: appsPaths) {
+                    details.appPath = config.getAppsPath() + i;
+                    appUsedKB += fs::getDirectorySpace(details.appPath);
+                }
+                details.appUsedKB = std::to_string(appUsedKB / 1024);
             }
             std::vector<std::string> dataPaths = dataBase->GetDataPaths(type, id);
             unsigned long long persistentUsedKB{};
@@ -384,11 +396,12 @@ uint32_t Executor::GetStorageDetails(const std::string& type,
                 details.persistentPath = config.getAppsStoragePath() + i;
                 persistentUsedKB += fs::getDirectorySpace(details.persistentPath);
             }
-            details.appUsedKB = std::to_string(appUsedKB / 1024);
             details.persistentUsedKB = std::to_string(persistentUsedKB / 1024);
+        } else {
+            return ERROR_WRONG_PARAMS;
         }
     } catch (std::exception& error) {
-        ERROR("Unable to retrive storage details: ", error.what());
+        ERROR("Unable to retrieve storage details: ", error.what());
         return Core::ERROR_GENERAL;
     }
     return ERROR_NONE;
@@ -402,7 +415,7 @@ uint32_t Executor::GetAppDetailsList(const std::string& type,
                           std::vector<DataStorage::AppDetails>& appsDetailsList) const
 {
     try {
-        appsDetailsList = dataBase->GetAppDetailsList(type, id, version, appName, category);
+        appsDetailsList = dataBase->GetAppDetailsListOuterJoin(type, id, version, appName, category);
     } catch (std::exception& error) {
         ERROR("Unable to get Applications details: ", error.what());
         return Core::ERROR_GENERAL;
@@ -437,6 +450,10 @@ uint32_t Executor::SetMetadata(const std::string& type,
                          const std::string& key,
                          const std::string& value)
 {
+    if (type.empty() || id.empty() || version.empty() || key.empty()) {
+        return ERROR_WRONG_PARAMS;
+    }
+
     try {
         dataBase->SetMetadata(type, id, version, key, value);
     } catch (std::exception& error) {
@@ -451,6 +468,10 @@ uint32_t Executor::ClearMetadata(const std::string& type,
                          const std::string& version,
                          const std::string& key)
 {
+    if (type.empty() || id.empty() || version.empty()) {
+        return ERROR_WRONG_PARAMS;
+    }
+
     try {
         dataBase->ClearMetadata(type, id, version, key);
     } catch (std::exception& error) {
@@ -465,6 +486,10 @@ uint32_t Executor::GetMetadata(const std::string& type,
                          const std::string& version,
                          DataStorage::AppMetadata& metadata) const
 {
+    if (type.empty() || id.empty() || version.empty()) {
+        return ERROR_WRONG_PARAMS;
+    }
+
     try {
         metadata = dataBase->GetMetadata(type, id, version);
     } catch (std::exception& error) {
@@ -658,14 +683,15 @@ void Executor::doUninstall(std::string type, std::string id, std::string version
 {
     INFO("type=", type, " id=", id, " version=", version, " uninstallType=", uninstallType);
 
-    dataBase->RemoveInstalledApp(type, id, version);
+    if (!version.empty()) {
+        dataBase->RemoveInstalledApp(type, id, version);
 
-    // TODO should "id" be also removed? what if two versions are installed?
-    auto appSubPath = Filesystem::createAppPath(id, version);
-    auto appPath = config.getAppsPath() + appSubPath;
+        auto appSubPath = Filesystem::createAppPath(id, version);
+        auto appPath = config.getAppsPath() + appSubPath;
 
-    INFO("removing ", appPath);
-    Filesystem::removeDirectory(appPath);
+        INFO("removing ", appPath);
+        Filesystem::removeDirectory(appPath);
+    }
 
     if (uninstallType == "full") {
         // only remove app record + storage when no other version installed
@@ -713,7 +739,7 @@ void Executor::doMaintenance()
             }
         }
 
-        auto appsDetailsList = dataBase->GetAppDetailsList();
+        auto appsDetailsList = dataBase->GetAppDetailsListOuterJoin();
         for (const auto& details : appsDetailsList) {
             INFO("details: ", details.id, ":", details.version);
